fix: #119 신청취소 모달 이슈
