+++ src/components/recruit/recruitDetail/footer/ApplierFooter.tsx
@@ -1,14 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { UnfilledBookmark, FilledBookmark } from '../../../../assets';
-import { useSetRecoilState, useRecoilState } from 'recoil';
+import { useSetRecoilState } from 'recoil';
 import { applyModalState, applyCancelModalState, needLoginModalState } from '../../../../atom';
 import { calculateDate } from '../../../../utils';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
-import { cancelApply } from '../../../../service';
 import { useBookmark, useLogin } from '../../../../hooks';
 import { useDelBookmark } from '../../../../hooks/useBookMark';
-import { NeedLogin } from '../../../index';
 
 interface ApplierData {
 	deadline: string;
@@ -22,15 +19,11 @@ const ApplierFooter = ({ deadline, isApplied, isBookmarked }: ApplierData) => {
 	const { isLoggedIn } = useLogin();
 	const setIsModal = useSetRecoilState(applyModalState);
 	const diffDate = calculateDate(deadline);
-	const queryClient = useQueryClient();
-	const [needLoginModal, setNeedLoginModal] = useRecoilState(needLoginModalState);
+	const setNeedLoginModal = useSetRecoilState(needLoginModalState);
 
 	const { mutate: bookmarked } = useBookmark({ queryKey: 'detailedPage' });
 	const { mutate: unBookmarked } = useDelBookmark({ queryKey: 'detailedPage' });
 
-	const cancelApplyTeam = useMutation({
-		mutationFn: (pageNum: number) => cancelApply(pageNum),
-	});
 	const setIsCancel = useSetRecoilState(applyCancelModalState);
 
 	const onClickApply = () => {
@@ -42,12 +35,7 @@ const ApplierFooter = ({ deadline, isApplied, isBookmarked }: ApplierData) => {
 	};
 
 	const onClickCancel = () => {
-		cancelApplyTeam.mutate(pageNum, {
-			onSuccess: () => {
-				setIsCancel(true);
-				queryClient.invalidateQueries({ queryKey: ['detailedPage'] });
-			},
-		});
+		setIsCancel(true);
 	};
 
 	const onClickBookmark = () => {
